Backend/src/utils/auth.py
"""
Auth helpers — JWT creation & password hashing.
"""

from __future__ import annotations

from datetime import datetime, timedelta, timezone

from jose import jwt
from passlib.context import CryptContext

from ..config import get_settings

_pwd_ctx = CryptContext(schemes=["bcrypt"], deprecated="auto")


class AuthService:
    """Handles password hashing and JWT token lifecycle."""

    def __init__(self) -> None:
        cfg = get_settings()
        self._secret = cfg.jwt_secret
        self._algorithm = cfg.jwt_algorithm
        self._expire_minutes = cfg.jwt_expire_minutes

    # ── passwords ──

    @staticmethod
    def hash_password(plain: str) -> str:
        return _pwd_ctx.hash(plain)

    @staticmethod
    def verify_password(plain: str, hashed: str) -> bool:
        return _pwd_ctx.verify(plain, hashed)

    # ── tokens ──

    def create_token(self, user_id: str, email: str) -> str:
        expire = datetime.now(timezone.utc) + timedelta(minutes=self._expire_minutes)
        payload = {"sub": user_id, "email": email, "exp": expire}
        return jwt.encode(payload, self._secret, algorithm=self._algorithm)

    def decode_token(self, token: str) -> dict:
        return jwt.decode(token, self._secret, algorithms=[self._algorithm])
