Backend/pyproject.toml
[tool.poetry]
name = "business-analysis-tool"
version = "0.1.0"
description = "Embedding + Contextual Graph based Business Analysis Tool - Cluster business points, embed documents with OpenAI, and visualize knowledge graphs"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://your-package-homepage.com"
repository = "https://github.com/yourusername/business-analysis-tool"
keywords = ["Business", "Analysis", "Clustering", "Visualization", "Embeddings", "Knowledge-Graph"]

[tool.poetry.dependencies]
python = "^3.11"

# --- Web Framework ---
fastapi = "^0.115.0"
uvicorn = {version = "^0.32.0", extras = ["standard"]}
python-multipart = "^0.0.12"

# --- Database & Vector Store ---
sqlalchemy = {version = "^2.0.36", extras = ["asyncio"]}
asyncpg = "^0.30.0"
pgvector = "^0.3.6"
alembic = "^1.14.0"
aiosqlite = "^0.20.0"

# --- OpenAI & LLM ---
openai = "^1.58.0"
tiktoken = "^0.8.0"

# --- LangChain & LangGraph ---
langchain = ">=0.3.0"
langchain-openai = ">=0.3.0"
langchain-community = ">=0.3.0"
langchain-experimental = ">=0.3.0"
langchain-text-splitters = ">=0.3.0"
langgraph = ">=0.2.0"

# --- Neo4j Knowledge Graph ---
neo4j = ">=5.25.0"

# --- Data Science & Clustering ---
pandas = "^2.2.0"
numpy = "^2.1.0"
scikit-learn = "^1.6.0"
scipy = "^1.14.0"

# --- Graph ---
networkx = "^3.4.0"

# --- Document Parsing ---
pypdf = "^5.1.0"
python-docx = "^1.1.0"
openpyxl = "^3.1.0"

# --- Utilities ---
pydantic = "^2.10.0"
pydantic-settings = "^2.7.0"
python-jose = {version = "^3.3.0", extras = ["cryptography"]}
passlib = {version = "^1.7.4", extras = ["bcrypt"]}
httpx = "^0.28.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.24.0"
black = "^24.0.0"
ruff = "^0.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
